Tidy ups and messy play arounds
extended a little bit more on entity and played around with the terrain.
Need to work with other ben on the state to get that shit sorted out. No
working unctionality for client as of yet, but will fix this asap

+++ src/server/TerrainManager.java
@@ -9,7 +9,7 @@ public class TerrainManager {
 	private SegmentGenerator sg = new SegmentGenerator(System.currentTimeMillis());
 	private Map<String, Segment> segmentCache = new HashMap<String, Segment>();
 	
-	public Segment loadSegment(long x, long y)
+	public Segment loadSegment(int x, int y)
 	{
 		String hash = Long.toString(x) + " " + Long.toString(y);
 		if(segmentCache.containsKey(hash))
